Refactor PdfCanvas component into specialized hooks
Major refactoring to improve code maintainability, testability, and separation of concerns:

## Changes Made:
- **usePdfRenderer hook** (308 lines): Extracted all PDF rendering logic
  - Single/spread page rendering
  - Canvas setup and scaling calculations
  - Error handling and render task management
  - Container size calculations

- **useCanvasInteraction hook** (173 lines): Extracted all mouse/touch interaction logic
  - Mouse drag, click, and wheel event handling
  - Page navigation logic for single/spread modes
  - Pan operations and zoom interactions
  - Global event management

- **Refactored PdfCanvas.tsx**: Reduced from 526 lines to 148 lines (72% reduction)
  - Clean separation of concerns
  - Simplified component logic using custom hooks
  - Maintained all existing functionality

## Benefits:
- **Maintainability**: Each hook has a single responsibility
- **Testability**: Smaller, focused units are easier to test
- **Reusability**: Hooks can be reused in other components
- **Readability**: Clear separation between rendering and interaction logic

## Quality Assurance:
- ✅ All 94 tests passing
- ✅ TypeScript compilation successful
- ✅ Production build verified
- ✅ No functional regressions
